+++ Changelog.md
@@ -5,7 +5,7 @@
     * This allows utilities like `syb:Data.Generics.Text.gread` to be meaningfully used at these types containing `ByteString`s.
   * [`fromListN` in `instance IsList ByteString` truncates input list if it's longer than the size hint](https://github.com/haskell/bytestring/pull/672)
 * API additions
-  * New TH splices: `Data.ByteString.thLiteral` and `Data.ByteString.thHexLiteral`
+  * New TH splices: `Data.ByteString.literalFromChar8` and `Data.ByteString.literalFromHex`
     * These validate input strings prior to generating corresponding
       compile-time literal ByteStrings.
 <!--